Solving PR review
