+++ src/examples/runPrintRhoSamples.py
@@ -1,10 +1,10 @@
 import sys
 
 
-import rhos.rhoUtils as u
-from rhos.rhoUtils import init_precision, LogMessage, end, Obs, adjust_precision, Inputs
-from rhos.rhoParser import parseArgumentPrintSamples
-from rhos import *
+import hltrho.utils.rhoUtils as u
+from hltrho.utils.rhoUtils import init_precision, LogMessage, end, Obs, adjust_precision, Inputs
+from hltrho.utils.rhoParser import parseArgumentPrintSamples
+from hltrho import *
 
 
 def init_variables(args_):
@@ -113,7 +113,7 @@ def main():
 
     for _e in range(par.Ne):
         estar_ = espace[_e]
-        fname = 'RhoSamplesE'+str(estar_)+'sig'+str(par.sigma)
+        fname = 'hltrhoamplesE'+str(estar_)+'sig'+str(par.sigma)
         fpath = os.path.join(par.logpath, fname)
         _Bnorm = (cNorm / (estar_ * estar_))
         _factor = (lambda_e[_e] * A0set[_e]) / _Bnorm
